Move classes out of package
